feat: add workspace button
